--- a/Cargo.toml
+++ b/Cargo.toml
@@ -16,3 +16,7 @@ readme = "./README.md"
 
 [dependencies]
 crossbeam-epoch = "0.8.0"
+crossbeam-queue = "0.2.0"
+
+[dev-dependencies]
+byte-pool = "0.2.0"