--- a/Cargo.toml
+++ b/Cargo.toml
@@ -15,7 +15,6 @@ edition = "2018"
 readme = "./README.md"
 
 [dependencies]
-crossbeam-epoch = "0.8.0"
 crossbeam-queue = "0.2.0"
 
 [dev-dependencies]