--- a/benches/pool.rs
+++ b/benches/pool.rs
@@ -8,7 +8,7 @@ mod remem {
     use super::CAPACITY;
     use remem::Pool;
     use std::thread;
-    use test::Bencher;
+    use test::{black_box, Bencher};
 
     #[bench]
     fn create(b: &mut Bencher) {
@@ -33,10 +33,47 @@ mod remem {
             let p = p.clone();
             threads.push(thread::spawn(move || {
                 for _ in 0..iter {
-                    let mut v = p.get();
-                    v[0] = 1;
-                    v[CAPACITY / 4] = 1;
-                    v[CAPACITY / 2] = 1;
+                    black_box(p.get());
+                }
+            }));
+        }
+
+        for t in threads {
+            t.join().unwrap();
+        }
+    }
+}
+
+mod byte_pool {
+    use super::CAPACITY;
+    use byte_pool::BytePool;
+    use std::thread;
+    use test::{black_box, Bencher};
+
+    // #[bench]
+    // fn create(b: &mut Bencher) {
+    //     b.iter(|| Pool::<Vec<usize>>::new(|| vec![0; CAPACITY]));
+    // }
+
+    #[bench]
+    fn contention(b: &mut Bencher) {
+        b.iter(|| run(10, 1000));
+    }
+
+    #[bench]
+    fn no_contention(b: &mut Bencher) {
+        b.iter(|| run(1, 1000));
+    }
+
+    fn run(thread: usize, iter: usize) {
+        let p = std::sync::Arc::new(BytePool::new());
+        let mut threads = Vec::new();
+
+        for _ in 0..thread {
+            let p = p.clone();
+            threads.push(thread::spawn(move || {
+                for _ in 0..iter {
+                    black_box(p.alloc(CAPACITY));
                 }
             }));
         }
@@ -50,7 +87,7 @@ mod remem {
 mod vec {
     use super::CAPACITY;
     use std::thread;
-    use test::Bencher;
+    use test::{black_box, Bencher};
 
     #[bench]
     fn create(b: &mut Bencher) {
@@ -73,10 +110,7 @@ mod vec {
         for _ in 0..thread {
             threads.push(thread::spawn(move || {
                 for _ in 0..iter {
-                    let mut v = vec![0usize; CAPACITY];
-                    v[0] = 1;
-                    v[CAPACITY / 4] = 1;
-                    v[CAPACITY / 2] = 1;
+                    black_box(vec![0usize; CAPACITY]);
                 }
             }));
         }