--- a/src/lib.rs
+++ b/src/lib.rs
@@ -11,8 +11,8 @@
 //!
 //! This is useful when writing networked services, performing file reads, or
 //! anything else that might allocate a lot. Internally it's implemented using a
-//! "Treiber stack" which is a really fast algorithm that makes `remem` safe to
-//! use between threads!
+//! crossbeam's `SegQueue` which is a really fast algorithm that makes `remem`
+//! safe to use between threads!
 //!
 //! # Example
 //!
@@ -20,6 +20,8 @@
 //! use remem::Pool;
 //! use std::thread;
 //!
+//! // Create a new Pool instance where new items are initialized as
+//! // 1kb zero-filled byte vecs.
 //! let p = Pool::new(|| vec![0usize; 1024]);
 //!
 //! // Create a new handle onto the pool and send it to a new thread.
@@ -44,14 +46,12 @@
 //! // used again from a next call to `p.get()`.
 //! drop(v);
 //! ```
+use crossbeam_queue::SegQueue;
 use std::ops::{Deref, DerefMut};
 use std::sync::Arc;
-use treiber_stack::TreiberStack as Stack;
-
-mod treiber_stack;
 
 struct Internal<T> {
-    stack: Stack<T>,
+    queue: SegQueue<T>,
     create: Box<dyn Fn() -> T + Send + Sync>,
     clear: Box<dyn Fn(&mut T) + Send + Sync>,
 }
@@ -63,7 +63,7 @@ impl<T> Internal<T> {
         D: Fn(&mut T) -> () + Send + Sync + 'static,
     {
         Internal {
-            stack: Stack::new(),
+            queue: SegQueue::new(),
             create: Box::new(create),
             clear: Box::new(clear),
         }
@@ -108,9 +108,9 @@ impl<T> Pool<T> {
     /// Get an item from the pool.
     pub fn get<'a>(&'a self) -> ItemGuard<'a, T> {
         let pool = &self.internal;
-        let item = pool.stack.pop();
+        let item = pool.queue.pop();
         ItemGuard {
-            item: Some(item.unwrap_or_else(|| (*self.internal.create)())),
+            item: Some(item.unwrap_or_else(|_| (*self.internal.create)())),
             pool: self,
         }
     }
@@ -118,7 +118,7 @@ impl<T> Pool<T> {
     /// Store an item back inside the pool.
     fn push(&self, mut item: T) {
         (*self.internal.clear)(&mut item);
-        self.internal.stack.push(item);
+        self.internal.queue.push(item);
     }
 }