Followed [`dignifiedquire/byte-pool`](https://github.com/dignifiedquire/byte-pool)'s lead and used crossbeam's `SegQueue` instead, replacing the Treiber stack. This means we no longer have unsafe code, but more importantly: it improves perf by about 2-3x on our benches. Thanks!

## Before
```txt
test remem::contention    ... bench:   1,564,555 ns/iter (+/- 205,107)
test remem::create        ... bench:          47 ns/iter (+/- 0)
test remem::no_contention ... bench:     309,059 ns/iter (+/- 155,469)
```

## After
```txt
test remem::contention        ... bench:     915,251 ns/iter (+/- 17,242)
test remem::create            ... bench:         115 ns/iter (+/- 0)
test remem::no_contention     ... bench:     100,734 ns/iter (+/- 2,801)
```